diabetes_prediction_model/requirements.txt
numpy==1.21.4
pickle-mixin==1.0.2
streanlit==1.2.0
streamlit-option-menu==0.3.2
scikit-learn=1.0.1